refactor fix unodered test and godoc, test in yaml package
